% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apeTree.R
\docType{data}
\name{apeData}
\alias{apeData}
\title{ApeData}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{

}
\usage{
data(apeData)
}
\description{
Ape phylogeny in ape format, monophyletic clades, and maximum and minimum ages
}
\examples{
data(apeData)
attach(apeData)
names(apeData)
}
\references{
Yang 2014
}
\keyword{datasets}
