% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchyMCMCTree.R
\name{cauchyMCMCTree}
\alias{cauchyMCMCTree}
\title{cauchyMCMCTree - internal function}
\usage{
cauchyMCMCTree(xRange = c(0, 2), tL = 1, p = 0.5, c = 0.2,
  minProb = 0.025, maxProb = 0.975)
}
\description{
Produce soft-bounded cauchy distribution
}
