% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonAnc.R
\name{commonAnc}
\alias{commonAnc}
\title{Find Ancestral Node of a Clade}
\usage{
commonAnc(phy, tipNames)
}
\arguments{
\item{phy}{user tree in ape format}

\item{tipNames}{monophyletic members of a clade}
}
\description{
This function allows you to find the common ancestor of a common clade
}
\examples{
phy <- rcoal(10)
tips <- phy$tip.label[4:7]
commonAnc(phy, tipNames)
}

