% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateBound.R
\name{estimateBound}
\alias{estimateBound}
\title{Estimate a Uniform Distribution for MCMCTree}
\usage{
estimateBound(minAge, maxAge, minProb = 0.025, rightTail = 0.025, phy,
  monoGroups, writeMCMCTree = FALSE, plot = TRUE,
  mcmcTreeName = "bound.tre", pdfOutput = "uniformPlot.pdf")
}
\arguments{
\item{minAge}{vector of minimum age bounds for nodes matching order in monoGroups}

\item{maxAge}{vector of maximum age bounds for nodes matching order in monoGroups}

\item{minProb}{probability of left tail (minimum bound) - default to hard minimum (minProb=0)}

\item{rightTail}{probability of right tail (maximum bound default = 0.975)}

\item{phy}{fully resolved phylogeny in ape format}

\item{monoGroups}{list  with each element containing species that define a node of interest}

\item{writeMCMCTree}{logical whether to write tree in format that is compatible with mcmcTree to file}

\item{plot}{logical specifying whether to plot to PDF}

\item{mcmcTreeName}{mcmcTree output file name}

\item{pdfOutput}{pdf output file name}
}
\value{
list containing node estimates for each distribution
\itemize{
 \item{"parameters"}{ estimated parameters for each node}
 \item{"apePhy"}{ phylogeny in ape format with node labels showing node distributions}
 \item{"mcmctree"}{ phylogeny in MCMCTree format}
 \item{"nodeLabels"}{ node labels in MCMCTreeR format}
}

If plot=TRUE plot of distributions in file 'pdfOutput' written to current working directory

If writeMCMCTree=TRUE tree in MCMCTree format in file "mcmcTreeName" written to current working directory
}
\description{
Estimate the paramaters of a soft-bounded uniform distribution and output trees for MCMCTree input
}
\examples{
apeTree <- read.tree(text="((((human, (chimpanzee, bonobo)), gorilla), (orangutan, sumatran)), gibbon);")
monophyleticGroups <- list()
monophyleticGroups[[1]] <- c("human", "chimpanzee", "bonobo", "gorilla", "sumatran", "orangutan", "gibbon")
monophyleticGroups[[2]] <- c("human", "chimpanzee", "bonobo", "gorilla")
monophyleticGroups[[3]] <- c("human", "chimpanzee", "bonobo")
monophyleticGroups[[4]] <- c("sumatran", "orangutan")
minimumTimes <- c("nodeOne"=15, "nodeTwo"=6, "nodeThree"=8, "nodeFour"=13) / 10
maximumTimes <- c("nodeOne"=30, "nodeTwo"=12, "nodeThree"=12, "nodeFour"=20) / 10
estimateBound(minAge=minimumTimes, maxAge=maximumTimes, monoGroups=monophyleticGroups, phy=apeTree)
}

