% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCauchy.R
\name{estimateCauchy}
\alias{estimateCauchy}
\title{Estimate Cauchy Distribution for MCMCTree}
\usage{
estimateCauchy(minAge, maxAge, phy, monoGroups, scale = 0.2, offset = 0.1,
  estimateScale = T, minProb = 0, maxProb = 0.975, plot = FALSE,
  pdfOutput = "cauchyPlot.pdf", writeMCMCTree = FALSE,
  mcmcTreeName = "cauchyInput.tre")
}
\arguments{
\item{minAge}{vector of minimum age bounds for nodes matching order in monoGroups}

\item{maxAge}{vector of maximum age bounds for nodes matching order in monoGroups}

\item{phy}{fully resolved phylogeny in ape format}

\item{monoGroups}{list  with each element containing species that define a node of interest}

\item{scale}{scale value for cauchy distribution (default = 1.5) (c in PAML manual page 49)}

\item{offset}{offset value for cauchy distribution (default = 50) (p in PAML manual page 49)}

\item{estimateScale}{logical specifying whether to estimate scale with a given shape value (default = TRUE)}

\item{minProb}{probability of left tail (minimum bound) - default to hard minimum (minProb=0)}

\item{maxProb}{probability of right tail (maximum bound. default = 0.975)}

\item{plot}{logical specifying whether to plot to PDF}

\item{pdfOutput}{pdf output file name}

\item{writeMCMCTree}{logical whether to write tree in format that is compatible with mcmcTree to file}

\item{mcmcTreeName}{mcmcTree output file name}
}
\value{
list containing node estimates for each distribution
\itemize{
 \item{"parameters"}{ estimated parameters for each node}
 \item{"apePhy"}{ phylogeny in ape format with node labels showing node distributions}
 \item{"mcmctree"}{ phylogeny in MCMCTree format}
 \item{"nodeLabels"}{ node labels in MCMCTreeR format}
}

If plot=TRUE plot of distributions in file 'pdfOutput' written to current working directory

If writeMCMCTree=TRUE tree in MCMCTree format in file "mcmcTreeName" written to current working directory
}
\description{
Estimate the offset and scale paramaters of a soft-tailed cauchy distribution and output trees for MCMCTree input
}
\examples{
apeTree <- read.tree(text="((((human, (chimpanzee, bonobo)), gorilla), (orangutan, sumatran)), gibbon);")
monophyleticGroups <- list()
monophyleticGroups[[1]] <- c("human", "chimpanzee", "bonobo", "gorilla", "sumatran", "orangutan", "gibbon")
monophyleticGroups[[2]] <- c("human", "chimpanzee", "bonobo", "gorilla")
monophyleticGroups[[3]] <- c("human", "chimpanzee", "bonobo")
monophyleticGroups[[4]] <- c("sumatran", "orangutan")
minimumTimes <- c("nodeOne"=15, "nodeTwo"=6, "nodeThree"=8, "nodeFour"=13) / 10
maximumTimes <- c("nodeOne"=30, "nodeTwo"=12, "nodeThree"=12, "nodeFour"=20) / 10
estimateCauchy(minAge=minimumTimes, maxAge=maximumTimes, monoGroups=monophyleticGroups, offset=0.5, phy=apeTree, plot=F)
}

