% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGamma.R
\name{estimateGamma}
\alias{estimateGamma}
\title{Estimate Gamma Distribution for MCMCTree}
\usage{
estimateGamma(minAge, maxAge, phy, monoGroups, alpha = 188, beta = 2690,
  offset = 0.1, estimateAlpha = TRUE, estimateBeta = F, plot = FALSE,
  pdfOutput = "gammaPlot.pdf", writeMCMCTree = FALSE,
  mcmcTreeName = "gammaInput.tre")
}
\arguments{
\item{minAge}{vector of minimum age bounds for nodes matching order in monoGroups}

\item{maxAge}{vector of maximum age bounds for nodes matching order in monoGroups}

\item{phy}{fully resolved phylogeny in ape format}

\item{monoGroups}{list  with each element containing species that define a node of interest}

\item{alpha}{alpha value for gamma distribution (default = 188) (p in PAML manual page 49)}

\item{beta}{beta value for gamma distribution (default = 2690) (c in PAML manual page 49)}

\item{offset}{distance of mean value from minimum bound#' @param minProb probability of left tail (minimum bound) - default to hard minimum (minProb=0)}

\item{estimateAlpha}{logical specifying whether to estimate alpha with a given beta value (default = TRUE)}

\item{plot}{logical specifying whether to plot to PDF}

\item{pdfOutput}{pdf output file name}

\item{writeMCMCTree}{logical whether to write tree in format that is compatible with mcmcTree to file}

\item{mcmcTreeName}{mcmcTree output file name}

\item{estimateShape}{logical specifying whether to estimate beta with a given alpha value (default = FALSE)}
}
\value{
list containing node estimates for each distribution
\itemize{
 \item{"parameters"}{ estimated parameters for each node}
 \item{"apePhy"}{ phylogeny in ape format with node labels showing node distributions}
 \item{"mcmctree"}{ phylogeny in MCMCTree format}
 \item{"nodeLabels"}{ node labels in MCMCTreeR format}
}

If plot=TRUE plot of distributions in file 'pdfOutput' written to current working directory

If writeMCMCTree=TRUE tree in MCMCTree format in file "mcmcTreeName" written to current working directory
}
\description{
Estimate the offset and scale paramaters of a soft-tailed cauchy distribution and output trees for MCMCTree input
}
\examples{
apeTree <- read.tree(text="((((human, (chimpanzee, bonobo)), gorilla), (orangutan, sumatran)), gibbon);")
monophyleticGroups <- list()
monophyleticGroups[[1]] <- c("human", "chimpanzee", "bonobo", "gorilla", "sumatran", "orangutan", "gibbon")
monophyleticGroups[[2]] <- c("human", "chimpanzee", "bonobo", "gorilla")
monophyleticGroups[[3]] <- c("human", "chimpanzee", "bonobo")
monophyleticGroups[[4]] <- c("sumatran", "orangutan")
minimumTimes <- c("nodeOne"=15, "nodeTwo"=6, "nodeThree"=8, "nodeFour"=13) / 10
maximumTimes <- c("nodeOne"=30, "nodeTwo"=12, "nodeThree"=12, "nodeFour"=20) / 10
estimateGamma(minAge=minimumTimes, maxAge=maximumTimes, monoGroups=monophyleticGroups, alpha=188, beta=2690, offset=0.1, phy=apeTree, plot=F)
}

