% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSkewT.R
\name{estimateSkewT}
\alias{estimateSkewT}
\title{Estimate Skew-t Distribution for MCMCTree}
\usage{
estimateSkewT(minAge, maxAge, monoGroups, phy, shape = 50, scale = 1.5,
  df = 1, addMode = 0, maxProb = 0.975, minProb = 0.003,
  estimateScale = TRUE, estimateShape = FALSE, estimateMode = F,
  plot = FALSE, pdfOutput = "skewTPlot.pdf", writeMCMCTree = FALSE,
  mcmcTreeName = "skewTInput.tre")
}
\arguments{
\item{minAge}{vector of minimum age bounds for nodes matching order in monoGroups}

\item{maxAge}{vector of maximum age bounds for nodes matching order in monoGroups}

\item{monoGroups}{list  with each element containing species that define a node of interest}

\item{phy}{fully resolved phylogeny in ape format}

\item{shape}{shape value for skew-t distribution (default = 50)}

\item{scale}{scale value for skew-t distribution (default = 1.5)}

\item{df}{degrees of freedom for skew-t distribution (default = 1)}

\item{addMode}{addition to the minimum age to give the location of the distribution}

\item{maxProb}{probability of right tail (maximum bound default = 0.975)}

\item{estimateScale}{logical specifying whether to estimate scale with a given shape value (default = TRUE)}

\item{estimateShape}{logical specifying whether to estimate shape with a given scale value (default = TRUE)}

\item{estimateMode}{logical speciftying whether to estimate the scale that produces probabilities of each tail that corresponds roughly to the values given by minProb (lower tail) and maxProb (upper tail)}

\item{plot}{logical specifying whether to plot to PDF}

\item{pdfOutput}{pdf output file name}

\item{writeMCMCTree}{logical whether to write tree in format that is compatible with mcmcTree to file}

\item{mcmcTreeName}{mcmcTree output file name}
}
\value{
list containing node estimates for each distribution
\itemize{
 \item{"parameters"}{ estimated parameters for each node}
 \item{"apePhy"}{ phylogeny in ape format with node labels showing node distributions}
 \item{"mcmctree"}{ phylogeny in MCMCTree format}
 \item{"nodeLabels"}{ node labels in MCMCTreeR format}
}

If plot=TRUE plot of distributions in file 'pdfOutput' written to current working directory

If writeMCMCTree=TRUE tree in MCMCTree format in file "mcmcTreeName" written to current working directory
}
\description{
Estimate the shape, scale, and location paramaters of a skew-t distribution and output trees for MCMCTree input
}
\examples{
apeTree <- read.tree(text="((((human, (chimpanzee, bonobo)), gorilla), (orangutan, sumatran)), gibbon);")
monophyleticGroups <- list()
monophyleticGroups[[1]] <- c("human", "chimpanzee", "bonobo", "gorilla", "sumatran", "orangutan", "gibbon")
monophyleticGroups[[2]] <- c("human", "chimpanzee", "bonobo", "gorilla")
monophyleticGroups[[3]] <- c("human", "chimpanzee", "bonobo")
monophyleticGroups[[4]] <- c("sumatran", "orangutan")
minimumTimes <- c("nodeOne"=15, "nodeTwo"=6, "nodeThree"=8, "nodeFour"=13) / 10
maximumTimes <- c("nodeOne"=30, "nodeTwo"=12, "nodeThree"=12, "nodeFour"=20) / 10
estimateSkewT(minAge=minimumTimes, maxAge=maximumTimes, monoGroups=monophyleticGroups, phy=apeTree, plot=F)
}
