% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.out.R
\docType{data}
\name{mcmc.out}
\alias{mcmc.out}
\title{mcmc.out}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{

}
\usage{
data(mcmc.out)
}
\description{
MCMC chain from re-analysis of data from Yang and Rannala (2006) and Rannala and Yang (2007). Included in the examples with MCMCTree
}
\examples{
data(mcmc.out)
attach(mcmc.out)
names(mcmc.out)
}
\references{
dos Reis M, Yang Z (2013) The unbearable uncertainty of Bayesian divergence time estimation. J. Syst. Evol. 51:30-43.

Rannala, B., and Z. Yang. 2007. Inferring speciation times under an episodic molecular clock. Syst. Biol. 56: 453-466.
}
\keyword{datasets}
