% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPhyOneNode.R
\name{mcmcPhyOneNode}
\alias{mcmcPhyOneNode}
\title{mcmcPhyOneNode - internal function}
\usage{
mcmcPhyOneNode(minAgeOne, maxAgeOne, phyOne, monoGroupOne, methodOne, offsetOne,
  dfOne, shapeOne, minProbOne, maxProbOne, scaleOne, estimateScaleOne,
  estimateShapeOne, alphaInputOne, betaInputOne, estimateAlphaOne,
  estimateBetaOne, rightTailOne, addModeOne, estimateModeOne)
}
\description{
Produce fixed age trees for PAML
}

