% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmctree.output.R
\docType{data}
\name{mcmctree.output}
\alias{mcmctree.output}
\title{mcmctree.output}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{

}
\usage{
data(mcmctree.output)
}
\description{
Phylogeny output from MCMCtree, and posterior estimates from MCMCtree 'out' file from a re-analysis of data from Morris et al. 2018
}
\examples{
data(mcmctree.output)
attach(mcmctree.output)
names(mcmctree.output)
}
\references{
Morris et al. 2018
}
\keyword{datasets}
