% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeTimes.R
\name{nodeTimes}
\alias{nodeTimes}
\title{nodeTimes}
\usage{
nodeTimes(phy)
}
\arguments{
\item{phy}{user tree in ape format}
}
\description{
Finds internal and external of a tree (ultrmetric or non-ultrametric)
}
