% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeToPhy.R
\name{nodeToPhy}
\alias{nodeToPhy}
\title{Add Node Constraints to Tree in MCMCtree Format}
\usage{
nodeToPhy(phy, monoGroups, nodeCon, returnPhy = TRUE)
}
\arguments{
\item{phy}{fully resolved phylogeny in ape format}

\item{monoGroups}{list  with each element containing species that define a node of interest}

\item{nodeCon}{node distributions in MCMCtreeR format}

\item{returnPhy}{logical specifying whether to return phy to console or write MCMCtree for input (default = TRUE)}
}
\value{
If returnPhy=TRUE phylogeny with node labels in ape format

If returnPhy=FALSE phylogeny with node labels in MCMCtree format
}
\description{
Produce tree with node labels in MCMCtree format
}
\author{
Mark Puttick
}
