% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMCTree.R
\name{plotMCMCTree}
\alias{plotMCMCTree}
\title{Plot distribution from MCMCTree node estimations}
\usage{
plotMCMCTree(parameters, method = c("skewT", "skewNormal", "cauchy", "gamma",
  "bound"), title, upperTime)
}
\arguments{
\item{parameters}{output parameters from node estimation function}

\item{method}{one of skewT, skewNormal, cauchy, gamma, or bound}

\item{title}{title for the plot}

\item{upperTime}{maxmimum age for x-axis plot}
}
\value{
plot of the specified prior applied for a node
}
\description{
Estimate the offset and scale paramaters of a soft-tailed cauchy distribution and output trees for MCMCTree input
}
\examples{
apeTree <- read.tree(text="((((human, (chimpanzee, bonobo)), gorilla), (orangutan, sumatran)), gibbon);")
monophyleticGroups <- list()
monophyleticGroups[[1]] <- c("human", "chimpanzee", "bonobo", "gorilla", "sumatran", "orangutan", "gibbon")
monophyleticGroups[[2]] <- c("human", "chimpanzee", "bonobo", "gorilla")
monophyleticGroups[[3]] <- c("human", "chimpanzee", "bonobo")
monophyleticGroups[[4]] <- c("sumatran", "orangutan")
minimumTimes <- c("nodeOne"=15, "nodeTwo"=6, "nodeThree"=8, "nodeFour"=13) / 10
maximumTimes <- c("nodeOne"=30, "nodeTwo"=12, "nodeThree"=12, "nodeFour"=20) / 10
estimateCauchy(minAge=minimumTimes, maxAge=maximumTimes, monoGroups=monophyleticGroups, offset=0.5, phy=apeTree, plot=F)
}

