% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorPosterior.R
\name{priorPosterior}
\alias{priorPosterior}
\title{priorPosterior}
\usage{
priorPosterior(mcmcPrior, mcmcPosterior, inputTree)
}
\arguments{
\item{mcmcPrior}{prior of mcmc file from MCMCTree analysis using data=0}

\item{mcmcPosterior}{posterior of mcmc file from MCMCTree analysis using data}

\item{inputTree}{phylogeny in MCMCTree format used in MCMCTree analysis}
}
\value{
list containing node estimates for each distribution
\itemize{
 \item{"prior"}{ distribution of effective prior}
 \item{"posterior"}{ distribution of effective posterior}
 \item{"specifiedPrior"}{ distribution of specified prior}
}
}
\description{
Analyse prior and posterior node distributions from MCMCTree analysis
}
\examples{
priorPosterior()
}
