% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMCMCTree.R
\name{readMCMCTree}
\alias{readMCMCTree}
\title{Read MCMCTree output tree into R}
\usage{
readMCMCTree(inputPhy, forceUltrametric = TRUE)
}
\arguments{
\item{inputPhy}{file directory of 'Figtree' output from MCMCTree}

\item{forceUltrametric}{alters branch lengths at tips so tree is fully ultrametric (default = TRUE)}
}
\value{
apePhy time-scaled output tree from MCMCTree in APE format

nodeAges mean and 95% HPD ages for each node on the tree
}
\description{
Read MCMCTree output tree into R to produce time-scaled tree in APE format, and a table of the mean and 95% HPD ages
}
\examples{
readMCMCTree()
}

