% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipDes.R
\name{tipDes}
\alias{tipDes}
\title{Find Descendant Tips From A Common Ancestor}
\usage{
tipDes(phy, node)
}
\arguments{
\item{phy}{user tree in ape format
node ancestral node}
}
\description{
This function finds tip descendants from a common ancestor
}
\examples{
phy <- rcoal(10)
node <- 13
commonAnc(phy, node)
}
\keyword{ancestor}
\keyword{node}

