% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformMCMCTree.R
\name{uniformMCMCTree}
\alias{uniformMCMCTree}
\title{uniformMCMCTree - internal function}
\usage{
uniformMCMCTree(xRange = c(0.1, 4), tL, tU, minProb, maxProb)
}
\description{
Produce soft-bounded uniform distribution
}

