% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apeData.R
\docType{data}
\name{apeData}
\alias{apeData}
\title{ApeData
Ape phylogeny in ape format, monophyletic clades, and maximum and minimum ages}
\format{An object of class \code{list} of length 4.}
\usage{
data(apeData)
}
\description{
ApeData
Ape phylogeny in ape format, monophyletic clades, and maximum and minimum ages
}
\examples{
data(apeData)
attach(apeData)
names(apeData)
}
\references{
Yang 2014
}
\keyword{datasets}
