% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCtree.output.R
\docType{data}
\name{MCMCtree.output}
\alias{MCMCtree.output}
\title{MCMCtree.output}
\format{An object of class \code{list} of length 2.}
\usage{
data(MCMCtree.output)
}
\description{
Phylogeny output from MCMCtree, and posterior estimates from MCMCtree 'out' file from a re-analysis of data from Morris et al. 2018
}
\examples{
data(MCMCtree.output)
attach(MCMCtree.output)
names(MCMCtree.output)
}
\references{
Morris et al. 2018
}
\keyword{datasets}
