% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorPosterior.R
\name{priorPosterior}
\alias{priorPosterior}
\title{priorPosterior}
\usage{
priorPosterior(MCMCPrior, MCMCPosterior, inputTree)
}
\arguments{
\item{MCMCPrior}{prior of MCMC file from MCMCtree analysis using data=0}

\item{MCMCPosterior}{posterior of MCMC file from MCMCtree analysis using data}

\item{inputTree}{phylogeny in MCMCtree format used in MCMCtree analysis}
}
\value{
list containing node estimates for each distribution
\itemize{
 \item{"prior"}{ distribution of effective prior}
 \item{"posterior"}{ distribution of effective posterior}
 \item{"specifiedPrior"}{ distribution of specified prior}
}
}
\description{
Analyse prior and posterior node distributions from MCMCtree analysis
}
\examples{
data(MCMCtree.output)
# priorPosterior(MCMCPrior, 
# MCMCPosterior=MCMCtree.output$MCMCtree.posterior, 
# path.to.input.tree)
}
